feat: allow setting initial theme from config/api
