+++ .changes/api-theme-window-option.md
@@ -1,0 +1,5 @@
+---
+"api": patch
+---
+
+Added `theme` field to `WindowOptions`.
